refactor(api): move controller from lib to src
