--- api/lib/application/passwords/password-controller.js
@@ -1,22 +1,0 @@
-import { usecases } from '../../domain/usecases/index.js';
-
-const updateExpiredPassword = async function (request, h) {
-  const passwordResetToken = request.payload.data.attributes['password-reset-token'];
-  const newPassword = request.payload.data.attributes['new-password'];
-  const login = await usecases.updateExpiredPassword({ passwordResetToken, newPassword });
-
-  return h
-    .response({
-      data: {
-        type: 'reset-expired-password-demands',
-        attributes: {
-          login,
-        },
-      },
-    })
-    .created();
-};
-
-const passwordController = { updateExpiredPassword };
-
-export { passwordController };
