+++ api/src/identity-access-management/application/password/password.controller.js
@@ -21,4 +21,21 @@ const createResetPasswordDemand = async function (request, h, dependencies = { r
   return h.response(serializedPayload).created();
 };
 
-export const passwordController = { checkResetDemand, createResetPasswordDemand };
+const updateExpiredPassword = async function (request, h) {
+  const passwordResetToken = request.payload.data.attributes['password-reset-token'];
+  const newPassword = request.payload.data.attributes['new-password'];
+  const login = await usecases.updateExpiredPassword({ passwordResetToken, newPassword });
+
+  return h
+    .response({
+      data: {
+        type: 'reset-expired-password-demands',
+        attributes: {
+          login,
+        },
+      },
+    })
+    .created();
+};
+
+export const passwordController = { checkResetDemand, createResetPasswordDemand, updateExpiredPassword };
