+++ api/tests/identity-access-management/unit/application/password/password.controller.test.js
@@ -23,7 +23,7 @@ describe('Unit | Identity Access Management | Application | Controller | passwor
       usecases.getUserByResetPasswordDemand.resolves({ email });
     });
 
-    it('should return serialized user', async function () {
+    it('returns serialized user', async function () {
       // when
       await passwordController.checkResetDemand(request, hFake, dependencies);
 
@@ -74,6 +74,43 @@ describe('Unit | Identity Access Management | Application | Controller | passwor
             type: 'password-reset-demands',
           },
         });
+      });
+    });
+  });
+
+  describe('#updateExpiredPassword', function () {
+    it('returns 201 http status code', async function () {
+      // given
+      const request = {
+        payload: {
+          data: {
+            attributes: {
+              'password-reset-token': 'PASSWORD_RESET_TOKEN',
+              'new-password': 'Password123',
+            },
+          },
+        },
+      };
+      sinon.stub(usecases, 'updateExpiredPassword');
+      usecases.updateExpiredPassword
+        .withArgs({
+          passwordResetToken: 'PASSWORD_RESET_TOKEN',
+          newPassword: 'Password123',
+        })
+        .resolves('beth.rave1221');
+
+      // when
+      const response = await passwordController.updateExpiredPassword(request, hFake);
+
+      // then
+      expect(response.statusCode).to.equal(201);
+      expect(response.source).to.deep.equal({
+        data: {
+          type: 'reset-expired-password-demands',
+          attributes: {
+            login: 'beth.rave1221',
+          },
+        },
       });
     });
   });
