--- api/tests/unit/application/passwords/password-controller_test.js
@@ -1,42 +1,0 @@
-import { passwordController } from '../../../../lib/application/passwords/password-controller.js';
-import { usecases } from '../../../../lib/domain/usecases/index.js';
-import { expect, hFake, sinon } from '../../../test-helper.js';
-
-describe('Unit | Controller | PasswordController', function () {
-  describe('#updateExpiredPassword', function () {
-    it('should return 201 http status code', async function () {
-      // given
-      const request = {
-        payload: {
-          data: {
-            attributes: {
-              'password-reset-token': 'PASSWORD_RESET_TOKEN',
-              'new-password': 'Password123',
-            },
-          },
-        },
-      };
-      sinon.stub(usecases, 'updateExpiredPassword');
-      usecases.updateExpiredPassword
-        .withArgs({
-          passwordResetToken: 'PASSWORD_RESET_TOKEN',
-          newPassword: 'Password123',
-        })
-        .resolves('beth.rave1221');
-
-      // when
-      const response = await passwordController.updateExpiredPassword(request, hFake);
-
-      // then
-      expect(response.statusCode).to.equal(201);
-      expect(response.source).to.deep.equal({
-        data: {
-          type: 'reset-expired-password-demands',
-          attributes: {
-            login: 'beth.rave1221',
-          },
-        },
-      });
-    });
-  });
-});
